LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/transformations/2d.transformation.scale.nonfinite-expected.txt

2d.transformation.scale.nonfinite

scale() with Infinity/NaN is ignored


PASS scale() with Infinity/NaN is ignored 
